only remove old backup if it exists

+++ backuphourly.sh
@@ -31,7 +31,9 @@ fi
 
 # check if there is a new hourly backup and output colorful success and error messages to ingame chat
 if [ -d "${backupdirectory}/hourly/${servername}-${newhourly}" ]; then
-	rm -r ${backupdirectory}/hourly/${servername}-${oldhourly}
+	if [ -d "${backupdirectory}/hourly/${servername}-${oldhourly}" ]; then
+		rm -r ${backupdirectory}/hourly/${servername}-${oldhourly}
+	fi
 	screen -Rd ${servername} -X stuff "tellraw @a [\"\",{\"text\":\"[Backup] \",\"color\":\"gray\",\"italic\":true},{\"text\":\"successfully created new backup\",\"color\":\"green\",\"italic\":true,\"hoverEvent\":{\"action\":\"show_text\",\"value\":{\"text\":\"\",\"extra\":[{\"text\":\"created file: ${servername}-${newhourly}, removed file: ${servername}-${oldhourly}, current world size: ${worldsize}\"}]}}}]$(printf '\r')"
 	echo "newest backup has been successfully created!" >> ${backuplog}
 	echo "added ${backupdirectory}/hourly/${servername}-${newhourly}" >> ${backuplog}
